improve term color support and list filters
Added some flags to update the list filters for some of the more
common operations that I use. Including listing only non-git dirs.
This forced me to refactor the info collecting code to allow
printouts to be more flexible and show the non-git dirs. Since I
was doing all of this anyways, I decided to add more colors and
proper color terminal detection.